renamed requestSubscriptions to requestService-client
